fine tuning

+++ src/main/java/com/jBrinkerExercise/JBexercise/Invoice.java
@@ -1,5 +1,7 @@
 package com.jBrinkerExercise.JBexercise;
 
+import java.math.BigDecimal;
+
 import org.json.JSONObject;
 
 public class Invoice extends JSONObject {
@@ -7,12 +9,12 @@ public class Invoice extends JSONObject {
 	private String ref;
 	private String date;
 	private String currency;
-	private double amount;
+	private BigDecimal amount;
 	
 	public Invoice() {
 	}
 
-	public Invoice(String ref, String date, String currency, double amount) {
+	public Invoice(String ref, String date, String currency, BigDecimal amount) {
 		super();
 		this.ref = ref;
 		this.date = date;
@@ -44,11 +46,11 @@ public class Invoice extends JSONObject {
 		this.currency = currency;
 	}
 
-	public double getAmount() {
+	public BigDecimal getAmount() {
 		return amount;
 	}
 
-	public void setAmount(double amount) {
+	public void setAmount(BigDecimal amount) {
 		this.amount = amount;
 	}
